Static File Service is Working
Text, JPG and HTML files are successfully served. If a file does not exist, a 404 code is sent. If a file exists, but does not match a supported file extension, a 501 code is sent. This behavior may not be ideal for security. Ther
e is a problem with parsing the file name using strtok where file names '.' and '..' are somehow ignored but '...'
is not. This may or may not be worth fixing.